cros-ide: Introduce fs wrapper for chroot and Chrome OS source
This patch introduces a service which abstracts away the difference
between inside and outside chroot and provides access to files under
chroot and Chrome OS source tree.
This patch doesn't contain any tests, but the next patch updates the
features to use the service and update the tests, and through which the
service will be tested.

BUG=b:232728571
TEST=npm t
TEST="Extension Tests" on VSCode running inside chroot

Change-Id: I5ffb55fec16a3dc7ea3519a974a089e59330578e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3650477
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>