portage_util: Disable check for GetReverseDeps
`equery depends ...` returns a status code of 1 for all packages that
are not installed in given the sysroot. cros_build_lib.run() throws
an exception for all status codes other than 0. We want to disregard
the results of `equery depends ...` when the package is not installed
so pass in 'check=False' to avoid throwing an exception in these
cases.

BUG=b:218521900
TEST=CQ, manual

Change-Id: Iddac1424fed4bd9bf86a546ce3ae9d0ddd66df42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3648687
Tested-by: Cindy Lin <xcl@google.com>
Commit-Queue: Ram Chandrasekar <rchandrasekar@google.com>
Auto-Submit: Cindy Lin <xcl@google.com>
Reviewed-by: Ram Chandrasekar <rchandrasekar@google.com>