cros-ide: Add a button to ignore error from C++ xrefs component
We have been dismissing error if the error has been shown at least once,
which means if user missed the error message, the message will never
been shown in the session. This patch updates the behavior so that the
message will reappear unless the user explicitly chose to ignore the
error message.

BUG=b:232043961
TEST=manually checked ignore button is added on error message, and
clicking it the same error is not shown anymore.

Change-Id: I8b740e1a61ab6c491da36205631367324eedf89a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3664175
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>