cros-ide: Generate compdb if the file has been removed
Previously we didn't regenerate compdb if it has been already generated
for the package, when we open C++ files. Especially we have not been
regenerating the compdb even if it has been removed manually.
This patch fixes the issue by making sure to regenerate the compdb if
the file is missing even if the extension has generated it before.

BUG=b:234398127
FIXED=b:234398127
TEST=npm t

Change-Id: Ic9c2438360fb21c0a63a31533d1bae98a35db375
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3686119
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>