cros_sdk: increase soft rlimit nproc to hard
On systems that set the nproc soft rlimit lower than the hard
limit, increase to the hard limit to mitigate cases where the
soft limit is too low to be usable.  This doesn't guarantee
the hard limit is enough, but there isn't much we can do if
the admin has set that too low.

BUG=b:234353695
TEST=CQ passes

Change-Id: I19e887864d8926fb04eb87a17e31fc6bc21617cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3686276
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Gilberto Contreras <gcontreras@google.com>