lib: Disable pylint checks for consider-using-with
Most use cases for the tempfile.{,Named}TemporaryFile wrapper methods in
this file are already written as with statement clauses, which would work
perfectly well for converting this method to use the @contextmanager
decorator to calm the linter.

Two notable exceptions are:
- In lib/parallel.py, the start() method explicitly sets the filestream
  for the _BackgroundTask class using these method calls, and the
  Cleanup() method then explicitly calls .close() on the file-like
  objects.
- In lib/cros_build_lib.py, run() indirectly calls these methods to set
  filestreams for stdout and stderr in some cases, and performs manual
  cleanup at the end of the run() function by calling .close() on these
  file-like objects. Similarly, run() also calls the built-in open()
  method in some cases, but also calls .close() at the end of the
  function.

These cases where filestreams are set explicitly are not wrapped in a
with statement, which makes a decorated generator function unusable
(since the underlying value is never accessed with the as keyword).
Since the tempfile methods being called here are functions and not
classes, a wrapper class cannot inherit from tempfile and must
re-implement accessors and modifiers for all required methods used by
file-like objects.

Given that the created resources are guaranteed to be cleaned up either
when .close() is explicitly called (as is the case in both usages noted
above) or, in the case of the calls to tempfile.{,Named}TemporaryFile,
when the objects are garbage-collected, I have opted to disable the
linter checks for the relevant lines instead. I'm happy to take a second
look if anyone objects.

BUG=b:234868086
TEST=./run_tests
TEST=cherrypick https://crrev.com/c/3636997 + cros lint

Change-Id: I263525d09a521ab9626e3770b983109785ab3aa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3697282
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Lizzy Presland <zland@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Lizzy Presland <zland@google.com>