cros-ide: Regenerate metrics config if user ID is set to @external
To resolve b/235428929, regenerate the metrics config is the user ID
is set to @external.

Also, this patch changes the metrics config to always contain a valid
UUID, and call isGoogler() in the metrics client instead to see if we
can upload metrics. This will save us from making further config tweaks
when we make changes to isGoogler() logic again.

BUG=b:235428929
TEST=npm test

Change-Id: Id367f72fdd89b620b3ec51e62d1e3a1ae1b6b6eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3699274
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>