ide_tooling: Add a script to update compdb to work outside chroot
Compilation database tells clangd how to compile the target C++ files.
Since the compilation database is generated inside chroot, we need to
make some modifications to it for it to be useful outside chroot.
This modification has been done in platform.eclass with sed and jq, but
it didn't have any tests.
In this patch we extract the logic to Python script and add unit tests
so that we can fix bugs without introducing regression.

See https://crrev.com/c/3705104 for platform.eclass change.

BUG=b:235426819
TEST=run the unit test

Change-Id: I9c62269725f3f2811463524430a91d9f090d0036
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3704981
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>