contrib: cl_perf: Handle edge case timestamps
It looks like it's possible (though rare) for a timestamp to not include
nanoseconds at the end. This leads to an IndexError on line 203. Example
timestamp is 2022-06-14T14:55:48Z.

Can workaround this by catching the exception and setting nanoseconds to
'0Z' but also have to remove the trailing 'Z' from the timestamp to
ensure proper formatting later.

BUG=None
TEST=manual, CQ

Change-Id: Ie4699e9f959443261751b50e3a84bfb16d940095
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3704805
Commit-Queue: Cindy Lin <xcl@google.com>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Tested-by: Cindy Lin <xcl@google.com>