cros_sdk: use user rlimits inside the SDK
When cros_sdk switches to root via `sudo`, the rlimits are set to
root's policies.  When we enter the chroot, we maintain those even
when we switch back to the user account.  Since all code run via
cros_sdk is effectively "user code", restore the user rlimits when
we enter the chroot.

This doesn't guarantee the limits are high enough for whatever the
SDK tries to do, but there isn't much we can do about that.  If the
system is configured with too low limits, the user and/or admin are
the ones who should resolve things.

BUG=b:234353695
TEST=`cros_sdk -- ulimit -a` reflects user settings instead of root

Change-Id: I0d08f725512e37e69a9d9257eb195073249b5681
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3721767
Reviewed-by: Ram Chandrasekar <rchandrasekar@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>