cros-ide: Stop using hard-coded mapping for C++ code completion
Before this patch, CrOS IDE was using a hard-coded mapping to get the
package name from the opened filepath for generating compdb. It means
this doesn't work for new packages, which is an issue. This patch
updates the mapping-generation logic so that the mapping is generated at
runtime. If the new algorithm misbehaves, the user can revert it to the
original hard-coded mapping from a setting.

BUG=b:237017032
TEST=npm t
TEST=manually checked several packages including cryptohome,
arc/key-master, and ml.

Change-Id: Idb970fc9acc10e9f5940e36742f89ebd38211c2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3720763
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>