service: Make differential linting more general.
Currently, differential linting is implemented for the CLI tool by
hardcoding the git repo path to check for modified lines of code.
In order to enable differential linting for other packages not located
in this hardcoded path, there needs to be a more generalized way of
determining the git repo path. This change adds a method for generating
that git repo path.

BUG=b:237430150
TEST=ran lint_packge to test the CLI tool

Change-Id: I54038bc4f7d3476245c3b720926fd97e2d369457
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3736045
Reviewed-by: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Uwem Wilson <uwemwilson@google.com>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: Uwem Wilson <uwemwilson@google.com>