service: fetch Golang lints produced by Golint.
Currently, the BuildLinter class can fetch clang-tidy and cargo clippy
lints. This change adds logic for fetching Golint lints that are
generated when a package using the cros-go eclass is built. In order to
support the addition of Golanf linting, it also modifies other chromite
files that use the BuildLinter class and the emerge_with_linting method.

BUG=b:236715258
TEST=ran lint_package CLI tool

Change-Id: Ia0210cdc5f354a91a134aef4a808bbc4c93bc0fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3723171
Commit-Queue: Uwem Wilson <uwemwilson@google.com>
Reviewed-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Tested-by: Uwem Wilson <uwemwilson@google.com>
Reviewed-by: Ryan Beltran <ryanbeltran@chromium.org>