lib: commandline: Add an "enum" action
When creating an action backed by an enum (e.g., --foo=some_value,
where Foo is an enum type, and some_value corresponds to an
enumeration of that type), one needs a lot of boilerplate code to add
the argument.

This adds an enum action which significantly simplifies the syntax of
doing so.

BUG=none
TEST=provided unit tests pass

Change-Id: I2d047d28d40eab00273de384f24bf4573a4dc155
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3749731
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>