utils: timer: Fix the timer delta printing as zero
timer.timer() context manager prints the time delta as 0 seconds. The
delta summary in the current context manager is printed before exiting
the context. The time delta is calculated only when we exit the context.
Hence the summary will print a delta of 0 always, which is the initial
default value.

The context manager is modified to print the summary after
exiting the Timer() context.

BUG=b:239234702
TEST=build_image

Change-Id: I17fbd7e5abe0b48a48328bf86ec88e9f1fbf499c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3766301
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Ram Chandrasekar <rchandrasekar@google.com>
Auto-Submit: Ram Chandrasekar <rchandrasekar@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>