cros-ide: Run compilation if necessary C++ files don't exist
We fix the issue that clangd fails for projects that use mojom bindings.

The updated algorithm goes as follows: we first use compdb_only USE flag
to generate compdb, and then parse the generated compdb and check
whether all the C++ files exist, and otherwise we consider those are
files to be generated, and run compilation using "compilation_database"
USE flag. Note that FEATURES=noclean is not relevant here, because we
don't run emerge, but we run ebuild command to run the phases up to
compile.

BUG=b:240373613
TEST=opened platform2/ml/benchmark.cc and confirmed code navigation
works for mojom flies.
TEST=opened platform2/codelab/main.cc and confirmed compilation phase
doesn't run as expected.

Change-Id: I30e0343c0b4b5220e94c89d6834929b680b5e48a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3792706
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>