lib: Provide standalone func for pkg size calc
In some cases, we want a library function to calculate a package's total
size when there is no Portage db instance on a partition. This change
allows a client to retrieve a package's file contents and separately
compute the size of the installed package.

BUG=b:224589938
TEST=unit

Change-Id: Ie43e48043d3cec1d9df76f69d30ebbdb05d5b3f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3805528
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Lizzy Presland <zland@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Lizzy Presland <zland@google.com>