cros-ide: Suggest user install Go for go linter.
Currently, if a user does not have the go binary installed and
able to be found by the `which go` command, the linter fails for
tast and shows an error. In order to make things clearer for
the user, this change suggests that the user install Go in order to
get proper Go linter functionality. This change also modifies the
LintConfig interface to make the executable an async function. This
is so that calling exec on `which go` can be awaited.

BUG=b:241845291
TEST=tested tast linter in environment with no go command in PATH
TEST=https://screenshot.googleplex.com/3RbZfNZgWByxt5Q

Change-Id: Ie264772652c8a93ab2c6fc1e0c8f1bb91fce8cbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3821748
Tested-by: Uwem Wilson <uwemwilson@google.com>
Commit-Queue: Uwem Wilson <uwemwilson@google.com>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Keigo Oka <oka@chromium.org>