cros_build_lib: run: drop RunCommandError.args property
I can't find any users of this as it represents the tuple of the
arguments that initialized the error.  The subprocess APIs don't
have anything equivalent to this, and when they do use "args",
it means "cmd" instead.  So drop it to avoid confusion.

BUG=b:187789262
TEST=CQ passes

Change-Id: Ida73e5e94a1ccc67e88c2ec9416de594a22b4e0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3837512
Reviewed-by: Sloan Johnson <sloanjohnson@google.com>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Sloan Johnson <sloanjohnson@google.com>