cli: chrome-sdk: Temporarily disable version caching
As part of rubik, we're moving the SDKs for public boards/configs to a
new GS bucket. When a user makes that switch (via the new
--use-new-public-bucket arg), part of the filesystem cache that stores
the 'full' version for an LKGM becomes invalidated. The buckets
use slightly different version formats.

So for the time being, we'll simply skip the caching for the version
string, and fetch it from GS at every `cros chrome-sdk` invocation.
The file is less than 20 bytes, so this shouldn't waste much time
at all.

We'll restore it once the new GS bucket has rolled out and there's
been a few LKGM updates afterward.

BUG=chromium:1330762
TEST=unittest
TEST=tested on chrome's CrOS bots: TODO

Change-Id: I22c2122a8fbd95c867654ffce04f45aad0a20508
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3843058
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>