scripts: downgrades ebuilds from EAPI 8 to 7
Adds a step at the end of the preparation phase to downgrade the ebuild
files for the target packages from EAPI 8 to 7. This is done by
replacing the number 8 with 7 in front of the "EAPI" keyword in the
ebuild file of the packages passed to the script for upgrade. This
enables "equery which" to detect the latest version of the target
package, which makes the rest of the automation possible.

Downgrading the EAPI can break the ebuild in some cases, as EAPI 7 may
not support all the features used in the file.

BUG=b:242758296
TEST=The script correctly updates the appropriate ebuild files, for the
specified packages only, and correctly identifies the latest versions.
TEST=If the file manipulation operation fails, the error is caught and
surfaced correctly.

Change-Id: I42263cd0aa0b92ecba80b172c7b8291aba204c36
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3842582
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>