lib: schedule SSP variant of autoupdate_EndToEndTest
The SSP version of this autotest has been ready for a while and seems
stable on per-build runs. Let's schedule it side-by-side for a bit to
make sure there are no problems when running in the real paygen
environment. If the results look OK after a few days, I'll switch it
over to only scheduling the SSP variant which will finish the conversion
work for this test.

BUG=b:228100799
TEST=./run_tests --debug lib/paygen

Change-Id: Iba2a2dbc1aae8e8e7781f999206ab3e21b40465a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3857491
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>