remote_access: change base_run to run
The run API will automatically init the remote work_dir when it's
needed.  The base_run API avoids that.  But in all existing cases,
the use of run suffices, so convert over so we can avoid exporting
multiple confusing APIs (which we'll do in a followup CL).

BUG=b:187791211
TEST=`./run_tests` passes

Change-Id: Iaacc7dd93841c300ab33e75f6455abdfa248620d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3860480
Reviewed-by: Sloan Johnson <sloanjohnson@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>