service: rework regen cache to avoid overhead
Make sure we calculate this content once, and log it once.  The current
code ends up logging the entire config file via the command line which
is extremely verbose when we update a lot of repos.

The binhost unittest had to be completely rewritten as it was largely
broken and not really testing anything at all.

BUG=b:244220188
TEST=`./bin/build_api ... chromite.api.BinhostService/RegenBuildCache` works

Change-Id: Ib6c4358e8272c83ccd5d75aea19b990cae93c92c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3868487
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ram Chandrasekar <rchandrasekar@google.com>