lib: add copy symbolic link
When running the tests in the DUT, it will copy the credentials, however
if the original credential is a symbolic link, it will only copy the
symbolic link. However, when it runs the tests, it will not be able to
access the symbolic link. So add "-L", it will copy the actual file.

BUG=None

Change-Id: I1505a74f750e800e213e883d6ae6d3307f88b641
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3880796
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Tested-by: Jieting Yang <yjt@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jieting Yang <yjt@google.com>