coverage: Update rust_coverage USE flag if set.
At present, we do not generate coverage for rust packages and we need to
use a different USE flag to generate code coverage for rust packages.
Accept a new field and add that USE flag when rust_code_coverage is set
in the request.

BUG=b:246409134
TEST=./run_tests service/test_unittests.py

Cq-Depend: chromium:3892792
Change-Id: Iabd4c2b6ff81ab06bd9aa404c484244ed2ae15d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3892108
Reviewed-by: Cindy Lin <xcl@google.com>
Tested-by: Srinivas Hegde <srinivashegde@google.com>
Commit-Queue: Srinivas Hegde <srinivashegde@google.com>