lib: toolchain_util: Handle kernel profiles from Arm
On Chrome we used chrome_cwp_profile to pass the arm arch.
We can't do the same trick with the kernel AFDO.
ArtifactProfileInfo can pass either kernel_version or
chrome_cwp_profile but not both.
This has to be fixed, b/246457583.

Meanwhile as a workaround we can prepend "arm-" to kernel_version to
distinguish the profiles.

BUG=b:244337204
TEST=unittests

Change-Id: I30e4f087ef1453b534297c5fa30a5c96065e509d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3893523
Tested-by: Denis Nikitin <denik@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Denis Nikitin <denik@chromium.org>