chrome_chromeos_lkgm: Rebase merge-conflict CLs if necessary
Once a new LKGM rolled, the existing uprev CLs needed to be rebased
since they have merge-conflict state. In this case, the gardeners
can manually rebase an existing CL if they want to reuse it, or they
can just abandon it if they want to wait for next LKGM roll oppotunity.

This CL introduces the automatic rebase in the former case so that they
don't need to do it manually. Even if the CL is not necessary, a
gardener can manually abandon the CL even after the automatic rebase.

BUG=b/245454643
TEST=`chrome_chromeos_lkgm_unittest` passes

Change-Id: I04fbc191461010059b7320343aefb9e7228d07b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3907129
Reviewed-by: Jack Neus <jackneus@google.com>
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Tested-by: Yoshiki Iguchi <yoshiki@chromium.org>