cros-ide: Remove install script
This patch removes the install script, which users don't need now that
the extension has been released to the marketplace.
This patch keeps the functionality to build and install an extension
from the source code for our development as a new shellscript file.

BUG=b:216583946
FIXED=b:216583946
TEST=install.sh is not mentioned in any document
TEST=manually tested `./dev_install.sh` and `./dev_install.sh --exe
code-server`

Change-Id: I663d8aa06ee819e4d2ac935d081625e7b9800e16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3919052
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>