OWNERS: Add crostc owners for toolchain things
This CL does a few things related to the ChromeOS Toolchains team in
Chromite OWNERS files. The goal is to make the crostc team code owners
for some files which we frequently modify, and also to clean up some
references to the ChromeOS Toolchains team wihin owners files. The
changes made are to:

1) Create OWNERS.crostc file to simplify references to the crostc team
2) Makes the crostc team code owners of lint_packages and related
   scripts
3) Removes an obsoleted reference to afdo stages for cbuildbot legacy
   builders

BUG=b:237711871
TEST=None

Change-Id: I5390b33c9b6efcb337daf71197c649061c5634e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3929122
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>