service: revision bump perfetto on only trunk refs
The git poller now sends trunk revisions or refs/tags/v* refs to
uprev_perfetto.

When the service receives only trunk refs (which is likely because
refs/tags/v* is created monthly), it assumes no version bump but only
a revision bump. When this happens, it uses the stable ebuild version as
the target version to bump the ebuild revision.

BUG=b:250513329
TEST=./run_tests service/packages_unittest.py lib/uprev_lib_unittest.py

Change-Id: Ia1f3a56df072ac0e7c465820707ed41f8e93856f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3934915
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Chinglin Yu <chinglinyu@chromium.org>