cros-ide: Add git directories watcher
This patch adds a watcher class which tracks visible git directories.
A git directory is defined to be visible if there is a visible text
editor managing a file under the directory.
The watcher fires an event when visible git directories change.

This class will be used by the spellchecker feature to spell check the
commit message for example, because the watcher allows the feature to
subscribe to the HEAD change event and rerun spell checker against the
HEAD commit message when it changes.

BUG=b:251746208
TEST=repo upload

Change-Id: Iba25c305882c88f12c34d2aa73ccaf78795b2746
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3951851
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>