lib: skyrim: update spi flash dut_control
A CPLD is no longer used to control the c2d2 SPI muxing. Skyrim now uses
a mux. The dut-control logic needs to be updated to account for
this.

Assert AP_FLASH_SELECT which controls access to the SPI flash. Revert the
sleep command to 5 seconds as we no longer need to wait for PWR_S5 to be
asserted by the EC (it was needed by the CPLD).

This change breaks the ability to flash Skyrim prototypes that have a
CPLD via c2d2.

BUG=b:254531120
TEST=`cros ap flash -b skyrim -i <image>` via c2d2 and servo

Change-Id: Ia3a9a868adb76509dfe2d53350f50cca94b3c67b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3969018
Tested-by: Mark Hasemeyer <markhas@google.com>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Mark Hasemeyer <markhas@google.com>
Auto-Submit: Mark Hasemeyer <markhas@google.com>