cros-ide: Warn if a chromiumos file is opened without a folder
This patch basically resurrect the handleNoChroot method which had been
added in https://crrev.com/c/3715397, but was removed in
https://crrev.com/c/3869039.

BUG=b:255446967
FIXED=b:255446967
TEST=repo upload
TEST=Open a codelab.cc without opening a folder in the workspace and
confirm an error message is shown. Then click the button in the error
message and confirm the platform2 directory is opened as a workspace
folder.

Change-Id: I21eb3511562b73b2c0cf53c694aad70cf2129135
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3977717
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>