cros-ide: add a setting to enable Gerrit comments
The new setting is used to guard whether Gerrit feature is enabled
(it is disabled by default).

We could have used the new setting inside gerrit.ts with
onDidChangeConfiguration to avoid the need for reload, but it is better
to have a kill switch in case of unforeseen problems.

BUG=b:216048068
TEST=manually enabled and disabled the feature

Change-Id: I85947195a7bd93cd60839876222525f5e1c9eef5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3977869
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>