cros-ide: Start Tast support under a flag
This patch creates the tast module and activates it when the flag is
enabled. We already have a run Tast tests feature under the
device_management module, which provides a command to run tast
tests under a DUT, but the new module is meant to support more general
Tast support such as providing the buttons to run tests, adding
debugging support, and updating the setting for providing code
completion and navigation.

BUG=b:255632861
TEST=repo upload
TEST=enable the flag and the message is shown.

Change-Id: Ic900c88416fa5c308acd45ddb08626e6422ddaf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3990227
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>