cros-ide: Automate set up for writing Tast tests
This patch adds a component to help developers writing Tast tests. As
the first feature of the component it checks VSCode is set up properly
so that code completion and navigation work, and otherwise automatically
sets up the environment with user confirmation.

BUG=b:255632861
TEST=repo upload
TEST=code completion and navigation works.
TEST=uninstall Go extension, open a Tast test file, and confirm an error
is shown. Click the button in the error message and confirm the Go
extension is installed.
TEST=remove go.gopath setting from the workspace JSON, open a Tast test
file, and confirm an error message is shown. Click the button in the
error message and confirm the gopath is set up and code navigation
works.
TEST=remove the tast directory from the workspace, open a Tast test
file, and confirm an error message is shown. Click the button in the
error message and confirm the tast directory is added to the workspace
and the window reloads. Then confirm code navigation works.

Change-Id: I846bc8e1e66dd49ef83bae781ca5543246815c2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3990228
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>