cros_mark_android_as_stable: deprecate --android_build_branch
The Android branch to use is pretty much hardcoded in the configuration
for each Android package; it's not like you can use a different branch
to uprev a package locally. Therefore there's little reason to make it a
input parameter.

This is a cleanup step in preparation for supporting more flexible
configuration (e.g. multiple packages pointing to the same branch).

BUG=b:255705023,b:255688351
TEST=run_tests
TEST=cros_mark_android_as_stable --android_package android-vm-rvc

Change-Id: Ib74c828fe52ccb27b6f727e197b7e745a382ca71
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3981875
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Shao-Chuan Lee <shaochuan@chromium.org>
Commit-Queue: Shao-Chuan Lee <shaochuan@chromium.org>