cros-ide: Full KB nav for Add device UI
This allows fast navigation of the UI for adding owned devices.
Everything can be done via the keyboard. Enter or Alt+ArrowRight
goes to the next step, Alt+ArrowLeft goes back a step, and on the
network type step Office and Home can be selected via the O and H
keys.

It was found that the window events were far more reliable and
straightforward than the HotKeys pkg and other similar pkgs.

BUG=b:249124164
TEST=manual

Change-Id: Id000c894b9912d8bf90e803cc17b96a1ee6facd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3984727
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Joel Becker <joelbecker@google.com>
Tested-by: Joel Becker <joelbecker@google.com>