scripts: make upload_prebuilts_unittest.py verify that options exist
http://crrev.com/c/3911178 introduced a new command-line option
named "output", but referred to it from the code as "json", resulting
in a run-time error. To avoid bugs like that in the future, this
change makes testMain in upload_prebuilts_unittest use a mock
options object specced by a real options object, so that only the
attributes that the real object has exist on the mock. This would
have caught the mistake during testing.

BUG=b:254910401
TEST=./run_tests; also tested that it catches the breaking change

Change-Id: Ie77dfc2639e9a8efdd3d5b4f29fbf397c7056d18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4011818
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Navil Perez <navil@google.com>
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Denis Nikitin <denik@chromium.org>