cbuildbot: Delete SDKPackageToolchainOverlaysStage
This class was only used by the SDK builder. The only usage was
commented out in 2016: https://crrev.com/c/332862

As we work on migrating the SDK builder to recipes, it is clear that
maintaining this stage is not a priority. Let's clean up the dead code.

BUG=b:250023328
TEST=./run_tests

Change-Id: Ifbfbdd9b270a8c71cbf6ecedd333bc779fcc5354
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4026506
Auto-Submit: Greg Edelston <gredelston@google.com>
Reviewed-by: Lizzy Presland <zland@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>
Tested-by: Greg Edelston <gredelston@google.com>