ide_tooling: Stop throwing error for missing trunk inside chroot
compdb_no_chroot.py has been asserting existence of the given file, but
it's actually not very meaningful because it runs inside chroot but
receives the file path to the chromiumos trunk outside chroot.
Usually the path also exists in chroot, but if the user renames the
trunk outside chroot, it is no longer true and it causes trouble.
This patch solves the problem by simply continuing the execution even
if the file path doesn't exist inside chroot.

BUG=b:259342928
FIXED=b:259342928
TEST=The user confirmed it resolved the issue http://b/259342928#comment2

Change-Id: I7a2d4a59bb1ecf5a930372ca16f6618951d33684
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4030078
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>