cros-ide: Update metrics to allow breakdown errors per user
We have been sending an exception hit for an error, but exception hits
don't allow us to see how many users are affected by a particular error
due to long-standing bug in GA. This patch updates metrics to
additionally send an event hit for an error so that we can see the
user counts in the dashboard.

BUG=b:259868838
TEST=repo upload
TEST=Install CrOS IDE with ./dev_install.sh, open
platform2/hermes/BUILD.gn, save the file, and confirm error happens and
the error event is reported in the dev dashboard.

Change-Id: I1798ccc8f0f3b4e648fde4378c7f0c52f89c4327
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4040803
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Eric Lok <lokeric@google.com>
Commit-Queue: Keigo Oka <oka@chromium.org>