service: Don't check include formats for iwyu list
This CL removes some assertions for IWYU include-list format which may
not pass for the final include list. It keeps these assertions for new
additions.

In the final suggested include list, IWYU will format new includes using
the previously enforced regex, but lines that are not changed may not follow that format. Furthermore we do not parse anything from this regex for these particular lines, so it should be safe to simply remove the check.

BUG=b:237320348
TEST=lint_package

Change-Id: I902f13c7639e3d37cb3d0abb97bdf13b9b5086b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4059989
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: Christopher Di Bella <cjdb@google.com>