cli: make cros clean should cros .cache directory instead of deleting
Changes the `cros clean` command so that it empties the chromiumos cache
directory instead of deleting it.

If the directory is deleted it breaks the mounting of the
`/var/cache/chromiumos-cache` directory in any chroots that happen to be
open until `cros_sdk` is run again and the cache directory is remounted.
While this mount is broken, `update_chroot` will fail.

BUG=b:242311131
TEST=Executed the following steps:
1) Create two shells, (S1 & S2).  In S1, enter the chroot, and make sure
there's some junk in the .cache and .cache/distfiles directories to
clean up.
2) In S2, run cros clean.
3) Back in S1, run update_chroot and verify that it completes.
4) Check that junk in .cache & .cache/distfiles has been deleted.

Change-Id: I84e1020cbb686f52573b4366d8b63bdb8230786d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4112729
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Phil Kiossoglou <pkiossoglou@chromium.org>
Commit-Queue: Phil Kiossoglou <pkiossoglou@chromium.org>
Auto-Submit: Phil Kiossoglou <pkiossoglou@chromium.org>