cros: usergroup_baseline: convert accounts to NamedTuple
This allows for typing information and to drop our non-standard class.

The uid/gid fields are mandatory, so even though we default them to
None, this is to maintain backwards compatibility.  We might be able
to fix this in the future by moving encpasswd to after those fields,
but that will require more source auditing first.

BUG=b:224601612
TEST=CQ passes

Change-Id: Ie1c92e83908ab4bdcb8e11b04660c66e7e255bb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4118367
Reviewed-by: Ram Chandrasekar <rchandrasekar@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>