cli: format/lint: ignore symlinks
It's not clear how we could reliably format or lint a symlinked file.
Do we read the file it's symlinked to, format it, and then write back
to the symlink target?  Or delete the symlink and format it with the
new data?

BUG=None
TEST=`cros lint` & `cros format` on symlinks get skipped

Change-Id: I4708586142fe6473d88bc67d905347073e61e664
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4121419
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>